feat: adding validation in configuration file

+++ src/index.ts
@@ -8,6 +8,7 @@ import { z } from "zod";
 import { getTemplate } from "@/helpers/getTemplate";
 import { handleError } from "@/utils/handleError";
 import { CONFIG_FILE_NAME, PKG_ROOT, PKG_TEMPLATE } from "@/consts";
+import { getConfig } from "./helpers/getConfig";
 
 process.on("SIGINT", handleError);
 process.on("SIGTERM", handleError);
@@ -50,11 +51,12 @@ async function main() {
     throw new Error("Invalid template. Please specify a valid url or path!");
   }
 
+  const config = await getConfig(isLocalProject ? cwd : PKG_ROOT);
+  if (!config)
+    throw new Error("Invalid configuration found. Please try again.");
   fs.removeSync(path.join(PKG_ROOT, CONFIG_FILE_NAME));
 
-  console.log(options);
-  console.log(isLocalProject);
-  console.log(cwd);
+  console.log(config);
 }
 
 main().catch(handleError);
